refactor(agent): satisfy the architecture ratchet this PR was breaking
Seven regressions, all from this branch — main passes clean. The ratchet is
explicit that the baseline records pre-existing debt, not exemptions for new
code, so six of the seven are fixed rather than absorbed.

- `agent -> schedule` was a new forbidden import: the snooze tool reached
  straight into SchedulerAPIClient. Wired through a composition adapter
  instead, mirroring `workflow_scheduler.ScheduleControlAdapter`, which
  exists for exactly this reason.
- `repositories.py` and `conversation_service.py` both grew past their
  oversized-file marks. The wait repository becomes `wait_repository.py`
  (its own aggregate, its own lifecycle) and the pause/resume primitive
  becomes `pause_resume.py` — which is what the branch's first commit
  claimed to be doing, now made literal. `_PAUSING_TOOL_NAMES` moves next
  to the code that consumes it so the two cannot drift.
- `handle_schedule_fired` hit complexity 16 (limit 15) with the snooze
  branch. Both wakes resume something already suspended and differ only in
  what they resume, so they collapse into one `_dispatch_wake`.

The seventh is a baseline bump: the sweep's per-item `except Exception` is
the correct shape for a batch that must not abort on one bad row, and is the
same pattern `RunResumeService.reconcile_stale_waits` already carries. Only
the two agent broad-catch counters move; nothing else is regenerated.

Adds a test for the composition adapter — after this refactor it is the only
place the real `schedule_once_job` signature is exercised.
