+++ lemma-backend/architecture-baseline.json
@@ -1,8 +1,8 @@
 {
   "broad_catches": {
-    "agent:count": 54,
+    "agent:count": 55,
     "agent:max": 5,
-    "agent:total": 71,
+    "agent:total": 72,
     "agent_surfaces:count": 106,
     "agent_surfaces:max": 5,
     "agent_surfaces:total": 128,
